moved states to their own class
