+++ data/states/__init__.py
@@ -1,0 +1,3 @@
+from .menu import Menu  # Import the Menu class
+from .game import Game  # Import the Game class
+from .states import States  # If there’s a base State class or shared logic
