+++ data/states/game.py
@@ -1,36 +1,11 @@
 import pygame as pg
-from .player import Player
+from .states import States
+from ..player import Player
 import data.tools as tools
-from .map import *
+from ..map import *
 from pygame.locals import *
 import data.constants as c  # Import constants
-import data.control as control
-  
-class States(object):
-    def __init__(self):
-        self.done = False
-        self.next = None
-        self.quit = False
-        self.previous = None
-  
-class Menu(States):
-    def __init__(self):
-        States.__init__(self)
-        self.next = 'game'
-    def cleanup(self):
-        print('cleaning up Menu state stuff')
-    def startup(self):
-        print('starting Menu state stuff')
-    def get_event(self, event):
-        if event.type == pg.KEYDOWN:
-            print('Menu State keydown')
-        elif event.type == pg.MOUSEBUTTONDOWN:
-            self.done = True
-    def update(self, screen, dt):
-        self.draw(screen)
-    def draw(self, screen):
-        screen.fill((255,0,0))
-  
+
 class Game(States):
     def __init__(self):
         States.__init__(self)
