+++ data/states/states.py
@@ -1,0 +1,7 @@
+class States(object):
+    def __init__(self):
+        self.done = False
+        self.next = None
+        self.quit = False
+        self.previous = None
+
